08/10/2024 Updates on Individual cycle inventories. Mkaer Checker validation for Accounts+Purchases, tocks and Cycle_Allocations table. Update on Recent Harvests table on dashboard page. Modified Purchase, stocks and Cycle_allocations table structures to have some columns in decimal format ->decimal('', 10, 2). Updated Checker pages to include Stocks, Purchases, Cycle_Allocations. .

+++ database/migrations/2024_10_02_152903_create_cycle_allocations_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->string('Cycle_Id');
             $table->unsignedBigInteger('stock_id');
-            $table->integer('allocated_quantity');
+            $table->decimal('allocated_quantity', 10, 2);
             $table->date('allocation_date');
-            $table->integer('remaining_quantity')->default(0); // Remaining quantity in this allocation (if partially used)
+            $table->decimal('remaining_quantity', 10, 2)->default(0); // Remaining quantity in this allocation (if partially used)
             $table->string('Status')->default('pending');
             $table->unsignedBigInteger('checker_id')->nullable();
             $table->unsignedBigInteger('maker_id')->nullable();
